layout: moved clear button to topbar
