+++ src/components/TopBar/Buttons.tsx
@@ -1,8 +1,10 @@
 import LightModeIcon from '@mui/icons-material/LightMode'
 import DarkModeIcon from '@mui/icons-material/DarkMode'
+import ClearIcon from '@mui/icons-material/Clear'
 import { useDispatch } from "react-redux"
 import { themeActions } from "../../redux/slices/themeSlice"
 import { IconButton, useTheme } from '@mui/material'
+import { outputActions } from '../../redux/slices/outputSlice'
 
 
 export const ToggleThemeButton = () => {
@@ -16,5 +18,22 @@ export const ToggleThemeButton = () => {
       {theme.palette.mode === 'light' ? <LightModeIcon /> : <DarkModeIcon />}
     </IconButton >
 
+  )
+}
+
+export const ClearButton = () => {
+  const dispatch = useDispatch()
+  const clearRequests = () => dispatch(outputActions.setRequests([]))
+  const clearResponses = () => dispatch(outputActions.setResponses([]))
+
+  return (
+    <IconButton
+      onClick={() => {
+        clearRequests()
+        clearResponses()
+      }}
+    >
+      <ClearIcon />
+    </IconButton>
   )
 }
